Update README.md
For the questions in this part, use the following

lists as needed:

mylist01 = [2,5,4,9,10,-3,5,5,3,-8,0,2,3,8,8,-2,-4,0,6]

mylist02 = [-7,-3,8,-5,-5,-2,4,6,7,5,9,10,2,13,-12,-4,1]

A1. Find the last four digits of the 13th Mersenne

prime, which is equal to 2^521 – 1.

A2. Find the product of the 7th entry in mylist01,

the 13th entry in mylist01, and the 4th entry

in mylist02.